sdk/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "trustmodel"
version = "0.1.0"
description = "TrustModel Agent Evaluation SDK - SSL/TLS for AI Agents"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "TrustModel", email = "team@trustmodel.dev" }
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Security",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
keywords = ["ai", "agents", "trust", "evaluation", "certificates", "safety"]

dependencies = [
    "httpx>=0.25.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp>=1.20.0",
    "pynacl>=1.5.0",
    "websockets>=12.0",
    "click>=8.1.0",
    "rich>=13.0",
    "structlog>=23.0",
    "tenacity>=8.2.0",
    "typing-extensions>=4.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.1",
    "mypy>=1.6",
    "ruff>=0.1",
    "respx>=0.20",
]
anthropic = ["anthropic>=0.18.0"]
openai = ["openai>=1.0"]
langchain = ["langchain>=0.1.0"]
all = [
    "anthropic>=0.18.0",
    "openai>=1.0",
    "langchain>=0.1.0",
]

[project.scripts]
trustmodel = "trustmodel.cli.main:cli"

[project.urls]
Homepage = "https://trustmodel.dev"
Documentation = "https://docs.trustmodel.dev"
Repository = "https://github.com/trustmodel/trustmodel-sdk"
Issues = "https://github.com/trustmodel/trustmodel-sdk/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/trustmodel"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
